+++ internal/services/azapi_resource_action_data_source_test.go
@@ -22,6 +22,30 @@ func TestAccActionDataSource_basic(t *testing.T) {
 	})
 }
 
+func TestAccActionDataSource_providerPermissions(t *testing.T) {
+	data := acceptance.BuildTestData(t, "azapi_resource_action", "test")
+	r := ActionDataSource{}
+
+	data.DataSourceTest(t, []resource.TestStep{
+		{
+			Config: r.providerPermissions(),
+			Check:  resource.ComposeTestCheckFunc(),
+		},
+	})
+}
+
+func TestAccActionDataSource_providerAction(t *testing.T) {
+	data := acceptance.BuildTestData(t, "azapi_resource_action", "test")
+	r := ActionDataSource{}
+
+	data.DataSourceTest(t, []resource.TestStep{
+		{
+			Config: r.providerAction(),
+			Check:  resource.ComposeTestCheckFunc(),
+		},
+	})
+}
+
 func (r ActionDataSource) basic(data acceptance.TestData) string {
 	return fmt.Sprintf(`
 %s
@@ -33,4 +57,35 @@ data "azapi_resource_action" "test" {
   response_export_values = ["*"]
 }
 `, GenericResource{}.defaultTag(data))
+}
+
+func (r ActionDataSource) providerPermissions() string {
+	return `
+provider "azurerm" {
+  features {}
+}
+
+data "azurerm_client_config" "current" {}
+
+data "azapi_resource_action" "test" {
+  type        = "Microsoft.Resources/providers@2021-04-01"
+  resource_id = "/subscriptions/${data.azurerm_client_config.current.subscription_id}/providers/Microsoft.Network"
+  action      = "providerPermissions"
+  method      = "GET"
+}
+`
+}
+
+func (r ActionDataSource) providerAction() string {
+	return `
+data "azapi_resource_action" "test" {
+  type        = "Microsoft.ResourceGraph@2020-04-01-preview"
+  resource_id = "/providers/Microsoft.ResourceGraph"
+  action      = "resources"
+  body = jsonencode({
+    query = "resources| limit 1"
+  })
+  response_export_values = ["*"]
+}
+`
 }