+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-## v1.8.0
+## v1.8.0 (unreleased)
 FEATURES:
 
 ENHANCEMENTS:
+- `azapi_resource_action`: Support provider action.
 
 BUG FIXES:
+- Fix a bug that resource id for type `Microsoft.Resources/providers` is not parsed correctly.
 - Fix a bug that resource id for type `Microsoft.Resources/tenants` is not parsed correctly.
 
 ## v1.7.0